+++ refinedstorage-platform-common/src/main/java/com/refinedmods/refinedstorage/platform/common/grid/CraftingGridContainerMenu.java
@@ -83,6 +83,7 @@ public class CraftingGridContainerMenu extends AbstractGridContainerMenu {
     public ItemStack quickMoveStack(final Player actor, final int slotIndex) {
         final Slot slot = getSlot(slotIndex);
         if (!actor.level().isClientSide() && slot instanceof CraftingGridResultSlot resultSlot) {
+            // TODO: check if craftedStack is not empty.
             final ItemStack craftedStack = resultSlot.onQuickCraft(actor);
             source.acceptQuickCraft(actor, craftedStack);
             return ItemStack.EMPTY;